LayoutTests/webdial/presence.html

<!DOCTYPE html>
<html>
<head>
    <script src="../resources/js-test-pre.js"></script>
</head>
<body>
    <script>
        description("Is WebDial present on navigator");
        shouldBeDefined(navigator.localNetworkDevices);
        shouldBeDefined(navigator.getRemoteInvoker);
        var devices = navigator.localNetworkDevices;
        if (devices) {
            shouldBeDefined(devices.enumerateDevices);
        }
    </script>
    <script src="../resources/js-test-post.js"></script>
</body>
</html>
